requirements.txt
wheel
flask
flask-login
flask-sqlalchemy
flask-babel
gunicorn